+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.6.4] 6-4-2018
 - Disabled query cache again as it's causing issues with offset/limit. Fixed [50](https://github.com/ClickSimply/Nano-SQL/issues/50).
+- Added optional constructor argument to LevelDB adapter for issue [49](https://github.com/ClickSimply/Nano-SQL/issues/49).
 
 ## [1.6.3] 5-27-2018
 - Using cache secondary index performance penalty is now only 10-50% depending on the adapter being used.